Fix asset bundle builds and tweak hoverfadetext to not have NRE
