+++ src/ValheimRAFT.Unity/.idea/.idea.ValheimRAFT.Unity/.idea/indexLayout.xml
@@ -2,7 +2,9 @@
 <project version="4">
   <component name="UserContentModel">
     <attachedFolders />
-    <explicitIncludes />
+    <explicitIncludes>
+      <Path>Assets/ValheimVehicles/SharedScripts/HoverFadeText.cs</Path>
+    </explicitIncludes>
     <explicitExcludes />
   </component>
 </project>
