+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 2953192193
+CRC: 3120327608
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
