+++ src/ValheimRAFT/AssetBundle/valheim-vehicles.manifest
@@ -1,15 +1,15 @@
 ManifestFileVersion: 0
-CRC: 1800126531
+CRC: 3294015059
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: 06e246b6d1697e8d2453399c7723b624
+    Hash: db94d8b8906c18393c277fa94a68a72a
   TypeTreeHash:
     serializedVersion: 2
-    Hash: b9588c41b767b224ab45a59ac1b63900
+    Hash: d8f024905b27cc255f2675d880be5359
   IncrementalBuildHash:
     serializedVersion: 2
-    Hash: 06e246b6d1697e8d2453399c7723b624
+    Hash: db94d8b8906c18393c277fa94a68a72a
 HashAppended: 0
 ClassTypes:
 - Class: 1
@@ -39,35 +39,35 @@ ClassTypes:
 - Class: 108
   Script: {instanceID: 0}
 - Class: 114
-  Script: {fileID: 11500000, guid: 9085046f02f69544eb97fd06b6048fe2, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 31a19414c41e5ae4aae2af33fee712f6, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 8233d90336aea43098adf6dbabd606a2, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 747419968f339df40a827088f25c215d, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: f44587a585f05c540ad46e160444cf0c, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 37593e1934874fe4db91a205982d109e, type: 3}
-- Class: 114
   Script: {fileID: 11500000, guid: 67db9e8f0e2ae9c40bc1e2b64352a6b4, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 1aa08ab6e0800fa44ae55d278d1423e3, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 2a4db7a114972834c8e4117be1d82ba3, type: 3}
 - Class: 114
   Script: {fileID: 11500000, guid: 0d0b652f32a2cc243917e4028fa0f046, type: 3}
 - Class: 114
-  Script: {fileID: 11500000, guid: 7d873279a1e81b9419d0f6fb8b3c1fa9, type: 3}
+  Script: {fileID: 11500000, guid: 5f7201a12d95ffc409449d95f23cf332, type: 3}
 - Class: 114
   Script: {fileID: 11500000, guid: 4e29b1a8efbd4b44bb3f3716e73f07ff, type: 3}
 - Class: 114
+  Script: {fileID: 11500000, guid: 1aa08ab6e0800fa44ae55d278d1423e3, type: 3}
+- Class: 114
   Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
 - Class: 114
-  Script: {fileID: 11500000, guid: 5f7201a12d95ffc409449d95f23cf332, type: 3}
+  Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 9085046f02f69544eb97fd06b6048fe2, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 747419968f339df40a827088f25c215d, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 31a19414c41e5ae4aae2af33fee712f6, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: f44587a585f05c540ad46e160444cf0c, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 8233d90336aea43098adf6dbabd606a2, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 37593e1934874fe4db91a205982d109e, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 2a4db7a114972834c8e4117be1d82ba3, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 7d873279a1e81b9419d0f6fb8b3c1fa9, type: 3}
 - Class: 115
   Script: {instanceID: 0}
 - Class: 120
@@ -194,8 +194,9 @@ Assets:
 - Assets/ValheimVehicles/Prefabs/shared_hull_slab_iron_4x4_new.prefab
 - Assets/ValheimVehicles/Prefabs/nautilus_bridge.prefab
 - Assets/ValheimVehicles/Prefabs/steering_wheel.prefab
-- Assets/ValheimVehicles/Prefabs/hull_wall_window_square_iron_2x2.prefab
+- Assets/ValheimVehicles/Prefabs/wheel_single_2.prefab
 - Assets/ValheimVehicles/Prefabs/hull_rib_wood.prefab
+- Assets/ValheimVehicles/Prefabs/hull_wall_window_square_iron_2x2.prefab
 - Assets/ValheimVehicles/Prefabs/GUI/edit_texture_panel 1.prefab
 - Assets/ValheimVehicles/Prefabs/hull_wall_wood_4x4.prefab
 - Assets/ValheimVehicles/Material/wood_alt_piece_bg_rib_8x4.mat
